Code split edit dialog (#2198)
* Code split edit dialog

* Code split edit view dialog
