+++ src/panels/lovelace/editor/hui-dialog-edit-view.ts
@@ -2,50 +2,21 @@ import { html, LitElement, PropertyDeclarations } from "@polymer/lit-element";
 import { TemplateResult } from "lit-html";
 
 import { HomeAssistant } from "../../../types";
-import { fireEvent, HASSDomEvent } from "../../../common/dom/fire_event";
-import { LovelaceViewConfig } from "../../../data/lovelace";
+import { HASSDomEvent } from "../../../common/dom/fire_event";
 import "./hui-edit-view";
 import "./hui-migrate-config";
+import { EditViewDialogParams } from "./show-edit-view-dialog";
 
 declare global {
   // for fire event
   interface HASSDomEvents {
     "reload-lovelace": undefined;
-    "show-edit-view": EditViewDialogParams;
   }
   // for add event listener
   interface HTMLElementEventMap {
     "reload-lovelace": HASSDomEvent<undefined>;
   }
 }
-
-let registeredDialog = false;
-const dialogShowEvent = "show-edit-view";
-const dialogTag = "hui-dialog-edit-view";
-
-export interface EditViewDialogParams {
-  viewConfig?: LovelaceViewConfig;
-  add?: boolean;
-  reloadLovelace: () => void;
-}
-
-const registerEditViewDialog = (element: HTMLElement) =>
-  fireEvent(element, "register-dialog", {
-    dialogShowEvent,
-    dialogTag,
-    dialogImport: () => import("./hui-dialog-edit-view"),
-  });
-
-export const showEditViewDialog = (
-  element: HTMLElement,
-  editViewDialogParams: EditViewDialogParams
-) => {
-  if (!registeredDialog) {
-    registeredDialog = true;
-    registerEditViewDialog(element);
-  }
-  fireEvent(element, dialogShowEvent, editViewDialogParams);
-};
 
 export class HuiDialogEditView extends LitElement {
   protected hass?: HomeAssistant;
@@ -98,4 +69,4 @@ declare global {
   }
 }
 
-customElements.define(dialogTag, HuiDialogEditView);
+customElements.define("hui-dialog-edit-view", HuiDialogEditView);
